INVOICE DELETE DELETE

+++ app/Modules/Crm/Http/Controllers/InvoiceController.php
@@ -95,7 +95,7 @@ class InvoiceController extends BaseController
             $invoice->invoice_no = $invNo;
             $invoice->order_id = isset($params['order_id']) ? $params['order_id'] : NULL;
             $invoice->store_id = $params['store_id'];
-            $invoice->customer_id = $customer_id =  $params['customer_id'];
+            $invoice->customer_id = $customer_id = $params['customer_id'];
             $invoice->discount_amount = 0;
             $invoice->grand_total = $grand_total = $params['grand_total'];
             $invoice->date = $date = $params['date'];
@@ -202,39 +202,6 @@ class InvoiceController extends BaseController
         return view('StoreInventory::brands.edit', compact('brands'));
     }
 
-    /**
-     * @param Request $request
-     * @return RedirectResponse
-     * @throws ValidationException
-     */
-    public function update(Request $request)
-    {
-        $this->validate($request, [
-            'name' => 'required|max:191',
-            'image' => 'mimes:jpg,jpeg,png|max:1000'
-        ]);
-        $params = $request->except('_token');
-        try {
-            $brand = Invoice::findOrFail($params['id']);
-            $collection = collect($params)->except('_token');
-            $logo = $brand->logo;
-            if ($collection->has('logo') && ($params['logo'] instanceof UploadedFile)) {
-                if ($brand->logo != null) {
-                    $this->deleteOne($brand->logo);
-                }
-                $logo = $this->uploadOne($params['logo'], 'brands');
-            }
-            $merge = $collection->merge(compact('logo'));
-            $brand->update($merge->all());
-
-            if (!$brand) {
-                return $this->responseRedirectBack('Error occurred while updating invoice.', 'error', true, true);
-            }
-            return $this->responseRedirect('crm.invoice.index', 'invoice updated successfully', 'success', false, false);
-        } catch (ModelNotFoundException $e) {
-            throw new ModelNotFoundException($e);
-        }
-    }
 
     /**
      * @param $id
@@ -243,21 +210,47 @@ class InvoiceController extends BaseController
     public function delete($id)
     {
         $data = Invoice::find($id);
-        $logo = $data->logo;
-        if ($data->delete()) {
-            if ($logo != null) {
-                $this->deleteOne($logo);
+        if ($data) {
+            $mr = DB::table('money_receipts')
+                ->where('invoice_id', '=', $id)->first();
+            if (!$mr) {
+                $inv_return = DB::table('invoice_returns')
+                    ->where('invoice_id', '=', $id)->first();
+                if (!$inv_return) {
+                    DB::table('inventories')->where('ref_id', $id)->where('ref_type', '=', Inventory::REF_INVOICE)->where('store_id', '=', $data->store_id)->delete();
+                    DB::table('invoice_details')->where('invoice_id', $id)->delete();
+                    if ($data->delete()) {
+                        return response()->json([
+                            'success' => true,
+                            'status_code' => 200,
+                            'message' => 'Invoice has been deleted successfully!',
+                        ]);
+                    } else {
+                        return response()->json([
+                            'success' => false,
+                            'status_code' => 200,
+                            'message' => 'Please try again!',
+                        ]);
+                    }
+                } else {
+                    return response()->json([
+                        'success' => false,
+                        'status_code' => 200,
+                        'message' => 'You cannot delete this invoice! Return Found on this invoice.',
+                    ]);
+                }
+            } else {
+                return response()->json([
+                    'success' => false,
+                    'status_code' => 200,
+                    'message' => 'You cannot delete this invoice! Money Receipt Found on this invoice.',
+                ]);
             }
-            return response()->json([
-                'success' => true,
-                'status_code' => 200,
-                'message' => 'Record has been deleted successfully!',
-            ]);
         } else {
             return response()->json([
                 'success' => false,
                 'status_code' => 200,
-                'message' => 'Please try again!',
+                'message' => 'Invoice Not Found!',
             ]);
         }
     }
